Disambiguate multiple QWaylandBuffer pointers
We carry multiple pointers to the same QWaylandBuffer object and when it
gets destroyed in a subclass, we still try to access the same object
from the superclass. Fixing this by removing the pointer from the
subclasses.
Also, to my understanding, we should signal the destruction to wayland by
a wl_destroy_buffer call from the client.

Change-Id: I67623e24452e92e9a4a0498b9b0b73d5b1b06c3a